app/requirements.txt
# CobbleAI App Server
anthropic==0.84.0
anyio==4.12.0
certifi==2025.4.26
charset-normalizer==3.4.2
clerk-backend-api==5.0.2
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Flask==3.1.3
flask-cors==6.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mcp==1.26.0
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.2
requests==2.32.3
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
Werkzeug==3.1.6
